requirements.txt
werkzeug==2.3.7
Pillow==9.5.0
easyocr==1.7.0
ImageHash==4.3.1
opencv-python==4.8.0.74
pyspellchecker==0.7.2
kombu==5.3.1
pytest==7.4.0
scenedetect[opencv]==0.6.2
pandas==2.0.3
apscheduler==3.10.4
numpy==1.26.4
tqdm==4.67.1
